🚚 entities and DTO for repository
