+++ backend/modules/onboarding/dto/__init__.py
@@ -1,0 +1,1 @@
+from .inputs import OnboardingUpdatableProperties
