+++ backend/modules/onboarding/entity/__init__.py
@@ -1,0 +1,1 @@
+from .onboarding import OnboardingStates
