+++ backend/modules/onboarding/repository/onboardings.py
@@ -1,39 +1,16 @@
 from datetime import datetime, timedelta
-from typing import Optional
-from uuid import UUID
 
 from fastapi import HTTPException
-from models.databases.repository import Repository
-from pydantic import BaseModel
+from modules.onboarding.entity.onboarding import OnboardingStates
+
+from .onboardings_interface import OnboardingInterface
 
 
-class OnboardingUpdatableProperties(BaseModel):
-
-    """Properties that can be received on onboarding update"""
-
-    onboarding_a: Optional[bool]
-    onboarding_b1: Optional[bool]
-    onboarding_b2: Optional[bool]
-    onboarding_b3: Optional[bool]
-
-    class Config:
-        extra = "forbid"
-
-
-class OnboardingStates(BaseModel):
-    """Response when getting onboarding"""
-
-    onboarding_a: bool
-    onboarding_b1: bool
-    onboarding_b2: bool
-    onboarding_b3: bool
-
-
-class Onboarding(Repository):
+class Onboarding(OnboardingInterface):
     def __init__(self, supabase_client):
         self.db = supabase_client
 
-    def get_user_onboarding(self, user_id: UUID) -> OnboardingStates | None:
+    def get_user_onboarding(self, user_id):
         """
         Get user onboarding information by user_id
         """
@@ -55,9 +32,7 @@ class Onboarding(Repository):
 
         return OnboardingStates(**onboarding_data[0])
 
-    def update_user_onboarding(
-        self, user_id: UUID, onboarding: OnboardingUpdatableProperties
-    ) -> OnboardingStates:
+    def update_user_onboarding(self, user_id, onboarding):
         """Update user onboarding information by user_id"""
         update_data = {
             key: value for key, value in onboarding.dict().items() if value is not None
@@ -76,7 +51,7 @@ class Onboarding(Repository):
 
         return OnboardingStates(**response[0])
 
-    def remove_user_onboarding(self, user_id: UUID) -> OnboardingStates | None:
+    def remove_user_onboarding(self, user_id):
         """
         Remove user onboarding information by user_id
         """
@@ -92,7 +67,7 @@ class Onboarding(Repository):
 
         return OnboardingStates(**onboarding_data[0])
 
-    def create_user_onboarding(self, user_id: UUID) -> OnboardingStates:
+    def create_user_onboarding(self, user_id):
         """
         Create user onboarding information by user_id
         """
@@ -110,7 +85,7 @@ class Onboarding(Repository):
 
         return OnboardingStates(**onboarding_data[0])
 
-    def remove_onboarding_more_than_x_days(self, days: int):
+    def remove_onboarding_more_than_x_days(self, days):
         """
         Remove onboarding if it is older than x days
         """
